src/workers/ocr-service/requirements.txt
requests==2.31.0
urllib3==2.0.7
python-json-logger==2.0.7
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python-headless>=4.10.0.84
Pillow>=10.4.0
pytest==8.1.1